+++ src/datamodel/clientDataModel.js
@@ -2,7 +2,7 @@ import dataUtils from "@/utils/dataUtils";
 
 export default {
     createResource(url, id, creationDate, title, description, language, paymentMethod, keywords, version, standardlicense,
-        publisher, fileType, policyNames, ruleIds, ruleJsons, artifactId, representationId, brokerUris, samples) {
+        publisher, fileType, policyNames, ruleIds, ruleJsons, artifactId, representationId, brokerUris, samples, catalogs) {
         let resource = {};
         if (url === undefined) {
             resource.url = "";
@@ -98,6 +98,11 @@ export default {
             resource.samples = [];
         } else {
             resource.samples = samples;
+        }
+        if (catalogs === undefined) {
+            resource.catalogs = [];
+        } else {
+            resource.catalogs = catalogs;
         }
         return resource;
     },
@@ -308,7 +313,7 @@ export default {
         return this.createApp(id, idsApp.title, idsApp.description, "APP");
     },
 
-    convertIdsResource(idsResource, representation, policyNames, ruleIds, ruleJsons, artifactId, brokerUris) {
+    convertIdsResource(idsResource, representation, policyNames, ruleIds, ruleJsons, artifactId, brokerUris, catalogs) {
         let title = idsResource.title;
         if (title.includes("\"@en")) {
             title = idsResource.title.substring(1, idsResource.title.lastIndexOf("\""));
@@ -326,7 +331,8 @@ export default {
 
         return this.createResource(idsResource._links.self.href, dataUtils.getIdOfConnectorResponse(idsResource), idsResource.creationDate, title, description,
             idsResource.language.replace("https://w3id.org/idsa/code/", ""), idsResource.paymentModality, idsResource.keywords,
-            idsResource.version, idsResource.license, idsResource.publisher, fileType, policyNames, ruleIds, ruleJsons, artifactId, representationId, brokerUris, idsResource.samples);
+            idsResource.version, idsResource.license, idsResource.publisher, fileType, policyNames, ruleIds, ruleJsons, artifactId, representationId, brokerUris,
+            idsResource.samples, catalogs);
     },