+++ src/pages/dataoffering/resources/addresource/AddResourcePage.js
@@ -3,6 +3,7 @@ import ResourceMetaDataPage from "./metadata/ResourceMetaDataPage.vue";
 import ResourcePolicyPage from "./policy/ResourcePolicyPage.vue";
 import ResourceRepresentationPage from "./representation/ResourceRepresentationPage.vue";
 import ResourceBrokersPage from "./brokers/ResourceBrokersPage.vue";
+import ResourceCatalogsPage from "./catalog/ResourceCatalogsPage.vue";
 import errorUtils from "../../../../utils/errorUtils";
 
 export default {
@@ -10,6 +11,7 @@ export default {
         ResourceMetaDataPage,
         ResourcePolicyPage,
         ResourceRepresentationPage,
+        ResourceCatalogsPage,
         ResourceBrokersPage
     },
     props: ['fromRoutePage'],
@@ -57,6 +59,10 @@ export default {
                     this.$refs.representationPage.gotVisible();
                 }
             } else if (this.$data.active_tab == 3) {
+                if (this.$refs.catalogsPage !== undefined) {
+                    this.$refs.catalogsPage.gotVisible();
+                }
+            } else if (this.$data.active_tab == 4) {
                 if (this.$refs.brokersPage !== undefined) {
                     this.$refs.brokersPage.gotVisible();
                 }
@@ -74,6 +80,7 @@ export default {
                 this.$refs.metaDataPage.loadResource(this.$data.currentResource, this.$data.onlyMetaData);
                 this.$refs.policyPage.loadResource(this.$data.currentResource);
                 this.$refs.representationPage.loadResource(this.$data.currentResource, false);
+                this.$refs.catalogsPage.loadResource(this.$data.currentResource);
                 this.$refs.brokersPage.loadResource(this.$data.currentResource);
             } catch (error) {
                 errorUtils.showError(error, "Get resource");
@@ -93,6 +100,7 @@ export default {
                 this.$refs.metaDataPage.loadResource(this.$data.currentResource, this.$data.onlyMetaData);
                 await this.$refs.policyPage.loadRequestedResource(this.$data.currentResource);
                 this.$refs.representationPage.loadResource(this.$data.currentResource, true);
+                this.$refs.catalogsPage.loadResource(this.$data.currentResource);
                 this.$refs.brokersPage.loadResource(this.$data.currentResource);
             } catch (error) {
                 errorUtils.showError(error, "Get resource");
@@ -108,6 +116,7 @@ export default {
             if (!onlyMetaData) {
                 this.$refs.policyPage.loadResource(this.$data.currentResource);
                 this.$refs.representationPage.loadResource(this.$data.currentResource, false);
+                this.$refs.catalogsPage.loadResource(this.$data.currentResource);
                 this.$refs.brokersPage.loadResource(this.$data.currentResource);
             }
             this.$data.active_tab = 0;
@@ -117,6 +126,7 @@ export default {
             this.$refs.metaDataPage.readonly = readonly;
             this.$refs.policyPage.readonly = readonly;
             this.$refs.representationPage.readonly = readonly;
+            this.$refs.catalogsPage.readonly = readonly;
             this.$refs.brokersPage.readonly = readonly;
         },
         async save() {
@@ -124,6 +134,7 @@ export default {
             if (this.$refs.representationPage.selected.length > 0) {
                 genericEndpoint = this.$refs.representationPage.selected[0];
             }
+
             var title = this.$refs.metaDataPage.title;
             var description = this.$refs.metaDataPage.description;
             var language = this.$refs.metaDataPage.language;
@@ -136,22 +147,24 @@ export default {
             var filetype = this.$refs.representationPage.filetype;
             var brokerList = this.$refs.brokersPage.getSelectedBrokerList()
             let brokerDeleteList = this.$refs.brokersPage.getBrokerDeleteList();
+            var catalogIds = this.$refs.catalogsPage.getSelectedCatalogsList();
+            let deletedCatalogIds = this.$refs.catalogsPage.getCatalogsDeleteList();
 
             if (this.fromRoutePage == 'true') {
                 // On route page this data is initially stored only in the node and will be saved with the route.
-                this.$emit("saved", title, description, language, paymentMethod, keywords, 0, standardlicense, publisher,
+                this.$emit("saved", catalogIds, title, description, language, paymentMethod, keywords, 0, standardlicense, publisher,
                     policyDescriptions, filetype, 0, brokerList);
             } else {
                 this.$root.$emit('showBusyIndicator', true);
                 this.$root.$emit('blockNavigationMenu', true);
                 if (this.$data.currentResource == null) {
-                    await dataUtils.createResourceWithMinimalRoute(title, description, language, paymentMethod, keywords, standardlicense, publisher,
+                    await dataUtils.createResourceWithMinimalRoute(catalogIds, title, description, language, paymentMethod, keywords, standardlicense, publisher,
                         policyDescriptions, filetype, brokerList, genericEndpoint);
                     this.$router.push('idresourcesoffering');
                     this.$root.$emit('showBusyIndicator', false);
                     this.$root.$emit('blockNavigationMenu', false);
                 } else {
-                    await dataUtils.editResource(this.$data.currentResource.id, this.$data.currentResource.representationId,
+                    await dataUtils.editResource(this.$data.currentResource.id, this.$data.currentResource.representationId, catalogIds, deletedCatalogIds,
                         title, description, language, paymentMethod, keywords, standardlicense, publisher, samples, policyDescriptions,
                         filetype, brokerList, brokerDeleteList, genericEndpoint, this.$data.currentResource.ruleId,
                         this.$data.currentResource.artifactId);