+++ src/pages/dataoffering/resources/addresource/catalog/dialog/AddCatalogDialog.vue
@@ -1,0 +1,3 @@
+<template src="./AddCatalogDialog.html"></template>
+<script src="./AddCatalogDialog.js"></script>
+<style src="./AddCatalogDialog.css"></style>